+++ example/settings.py
@@ -147,7 +147,8 @@ PROVENANCE = {
     },
     "OTHER": {
         # sets the max recorded length of an argument that is passed into a recorded function
-        "MAX_ARG_LENGTH": 100
+        "MAX_ARG_LENGTH": 100,
+        "MAX_FIELD_VALUE_LENGTH": 50
     }
 
 }